PR(UNRELATED): Make a fixture for existing 1-to-m
